source/requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.3
mdurl==0.1.2
narwhals==1.40.0
numpy==1.26.4
packaging==23.2
pandas==2.2.3
pillow==10.4.0
protobuf==4.25.7
pyarrow==20.0.0
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.25.1
six==1.17.0
smmap==5.0.2
streamlit==1.32.0
tenacity==8.5.0
toml==0.10.2
tornado==6.5.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
